refactor: rename component CartItemsList to CartProductsList
