+++ src/widgets/cart-products-list/lib/helpers.ts
