+++ src/widgets/cart-products-list/model/hooks.ts
