+++ src/widgets/cart-products-list/ui/CartProductsList.tsx
